requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
darkdetect==0.8.0
enum-compat==0.0.3
idna==3.10
iniconfig==2.1.0
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pluggy==1.6.0
pygments==2.19.2
pyqt6==6.9.1
pyqt6-fluent-widgets==1.9.0
pyqt6-frameless-window==0.7.3
pyqt6-qt6==6.9.2
pyqt6-sip==13.10.2
pytest==8.4.2
pyueye==4.96.952
pywin32==311
pyyaml==6.0.2
refractiveindex2==0.1.0
requests==2.32.5
scipy==1.16.2
types-pyyaml==6.0.12.20250915
types-requests==2.32.4.20250913
urllib3==2.5.0
